+++ src/components/Button.tsx
@@ -1,3 +1,5 @@
 import Button from 'reactstrap/lib/Button';
 
+export type { ButtonProps } from 'reactstrap/lib/Button';
+
 export default Button;
